Fix020: Fixed Portfolio now  Still an error on Skill setting linking pages

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\DataTables\CategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\PortfolioItem;
 use Illuminate\Http\Request;
 
 class CategoryController extends Controller
@@ -94,9 +95,14 @@ class CategoryController extends Controller
     {
         //
         $category = Category::findOrFail($id);
-        $category->delete();
-
-        toastr()-> success('Category deleted successfully');
+        $hasItem = PortfolioItem::where('category_id', $category->id)->count();
+        if ($hasItem == 0) {
+            $category->delete();  
+            return true;   
+        }
+        
+        return response(['status' => 'error']);
+        //toastr()-> success('Category deleted successfully');
 
         
         // return redirect()->route('admin.category.index');
